requirements-dev.txt
# ========================================
# ZINA Cantine BAD - Development Dependencies
# ========================================

# Testing Framework
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0
pytest-asyncio==0.21.1
pytest-xdist==3.3.1

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
pylint==3.0.3
bandit==1.7.5

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
mkdocs==1.5.3
mkdocs-material==9.4.8

# Development Tools
pre-commit==3.5.0
python-dotenv==1.0.0
watchdog==3.0.0
rich==13.7.0

# Database Tools
alembic==1.12.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# API Documentation
flask-restx==1.3.0
marshmallow==3.20.1
apispec==6.3.0

# Performance Profiling
py-spy==0.3.14
memory-profiler==0.61.0
line-profiler==4.1.1

# Security
safety==2.3.5
pip-audit==2.6.1
bandit==1.7.5

# Environment Management
pipenv==2023.11.15
virtualenv==20.24.7

# Linting and Formatting
autopep8==2.0.4
yapf==0.40.2
docformatter==1.7.3

# Type Checking
types-requests==2.31.0.10
types-python-dateutil==2.8.19.14
types-PyYAML==6.0.12.12

# Debugging
ipdb==0.13.13
pdb++==0.10.3
icecream==2.1.3

# Task Management
invoke==2.2.0
fabric==3.2.2

# Data Validation
cerberus==1.3.5
jsonschema==4.20.0
pydantic==2.5.0

# Monitoring and Logging
structlog==23.2.0
sentry-sdk==1.38.0
prometheus-client==0.19.0

# Migration Tools
flask-migrate==4.0.5
flask-script==2.0.6

# Email Testing
email-validator==2.1.0
aiosmtplib==3.0.1

# Cache Testing
redis==5.0.1
fakeredis==2.20.1

# HTTP Testing
httpx==0.25.2
requests-mock==1.11.0
responses==0.23.3

# Time Testing
freezegun==1.2.2
time-machine==2.13.0

# File Handling
python-magic==0.4.27
pillow==10.1.0

# Background Jobs
celery==5.3.4
kombu==5.3.4

# WebSocket Testing
websockets==12.0

# GraphQL Testing
graphene==3.3
graphql-core==3.2.3

# Mobile Testing
appium-python-client==2.11.1

# Load Testing
locust==2.17.0

# Container Tools
docker==6.1.3
docker-compose==1.29.2

# Cloud Tools
boto3==1.34.0
google-cloud-storage==2.10.0
azure-storage-blob==12.19.0

# Data Analysis
pandas==2.1.4
numpy==1.26.2

# Machine Learning (for recommendations)
scikit-learn==1.3.2
tensorflow==2.15.0

# Natural Language Processing
nltk==3.8.1
spacy==3.7.2

# Image Processing
opencv-python==4.8.1.78
scikit-image==0.22.0

# PDF Generation
reportlab==4.0.7
weasyprint==60.2

# Excel Generation
openpyxl==3.1.2
xlsxwriter==3.1.9

# QR Code Generation
qrcode[pil]==7.4.2
pyzbar==0.1.9

# SMS Integration
twilio==8.10.3
africastalking==1.2.7

# Payment Integration
stripe==7.5.0
paypal-checkout-serversdk==1.0.3

# Social Media Integration
python-facebook-api==0.17.1
tweepy==4.14.0

# SEO Tools
beautifulsoup4==4.12.2
selenium==4.15.2

# Internationalization
babel==2.13.1
polyglot==16.7.4

# Configuration Management
dynaconf==3.2.4
configparser==6.0.0

# Template Testing
jinja2==3.1.2
mako==1.3.0

# WebSocket Support
flask-socketio==5.3.6
python-socketio==5.10.0

# Task Queues
rq==1.15.1
dramatiq==1.15.0

# Rate Limiting
flask-limiter==3.5.0
redis-rate-limit==1.3.1

# Caching
flask-caching==2.1.0
diskcache==5.6.3

# Session Management
flask-session==0.5.0
redis-session==0.4.1

# API Versioning
flask-restplus==0.13.0
flask-smorest==0.44.0

# Data Export
pandas==2.1.4
openpyxl==3.1.2

# Chart Generation
matplotlib==3.8.2
plotly==5.17.0
seaborn==0.13.0

# Map Integration
folium==0.15.0
geopy==2.4.1

# Weather Integration
pyowm==3.3.0
meteostat==1.6.7

# Notification Systems
notifier==1.0.3
plyer==2.1.0

# Backup Tools
duplicity==1.3.2
borgbackup==1.2.4

# Health Checks
healthcheck==1.3.3
flask-healthz==0.1.0
